clear cart after payment, escape transaction cors
