+++ server/functions/controllers/TransactionController.js
@@ -102,6 +102,7 @@ const createTransaction = async (req, res, next) => {
             }
 
             const useProxy = boolean(process.env.USE_PROXY);
+
             let response = '';
             let params = '';
 
@@ -287,41 +288,6 @@ const checkPayment = async (refId) => {
     }
 }
 
-const getTransaction = async (req, res) => {
-    const { idTransaction } = req.params;
-
-    const transaction = await TransactionModel.findOne({ refId: idTransaction });
-    if (!transaction) {
-        return res
-            .status(404)
-            .json({ success: false, msg: `Transaction not found` });
-    }
-
-    const paymentMethodName = Object.keys(transaction.paymentMethod)[0];
-
-    if (transaction.paymentMethod[paymentMethodName].status === 'PENDING') {
-        await checkPayment(idTransaction);
-    }
-
-    const restaurant = await RestaurantModel.findById(transaction.idRestaurant).select("idOwner name address").exec();
-
-    if (!restaurant) {
-        return res
-            .status(404)
-            .json({ success: false, msg: `Restaurant not found` });
-    }
-
-    const merchant = await MerchantModel.findById(restaurant.idOwner).select("ico").exec();
-
-    if (!merchant) {
-        return res
-            .status(404)
-            .json({ success: false, msg: `Merchant not found` });
-    }
-
-    return res.status(200).json({ success: true, msg: { transaction: transaction, restaurant: restaurant, merchant: merchant } });
-}
-
 const getPaymentMethods = async (req, res) => {
     const { idRestaurant } = req.params;
 
@@ -355,7 +321,7 @@ const getPaymentMethods = async (req, res) => {
 }
 module.exports = {
     createTransaction,
-    getTransaction,
     getPaymentMethods,
-    runAutoCheckPayment
+    runAutoCheckPayment,
+    checkPayment
 }
