+++ server/functions/routes/TransactionRouter.js
@@ -5,10 +5,7 @@ const PAYMENT_METHODS = require('../config/api').PAYMENT_METHODS;
 const router = express.Router();
 const authMiddleware = require('./AuthMiddlewares');
 
-
 router.post('/' + TRANSACTION, TransactionController.createTransaction, authMiddleware.validationHandlerMiddleware);
-
-router.get('/' + TRANSACTION + '/:idTransaction', TransactionController.getTransaction);
 router.get('/' + TRANSACTION + '/' + PAYMENT_METHODS + '/:idRestaurant', TransactionController.getPaymentMethods);
 
 module.exports = router;
